requirements.txt
langchain
llama-index
chromadb
duckduckgo-search
langgraph
beautifulsoup4
fastapi
uvicorn
python-dotenv